app/requirements.txt
fastapi
uvicorn[standart]
sqlalchemy
passlib[bcrypt]
alembic
python-dotenv
pydantic[email]
python-multipart
aiofiles
pytest
redis
pyjwt[crypto]
pytest-asyncio
httpx
fastapi-babel
python-slugify
factory-boy
minio
Pillow
pyotp
websockets
sentry-sdk[fastapi]
numpy
celery
flower
psycopg2-binary
fastapi-sqlalchemy
gunicorn
python-jose